@peppy 2022-09-15
I doubt it's going to be helpful in any way, but initially I tried a recursive implementation of this. It works but code readability seems worse and it only avoids a few duplicated calls, so I wouldn't see it being hugely beneficial without thousands of types.

Here's the diff before I throw away the branch:

```diff
diff --git a/osu.Framework/Graphics/Drawable.cs b/osu.Framework/Graphics/Drawable.cs
index 6a54c2116..501d53cfc 100644
--- a/osu.Framework/Graphics/Drawable.cs
+++ b/osu.Framework/Graphics/Drawable.cs
@@ -132,48 +132,41 @@ protected virtual void Dispose(bool isDisposing)
         /// </summary>
         internal virtual void UnbindAllBindablesSubTree() => UnbindAllBindables();
 
-        private void cacheUnbindActions()
+        private Action<object> cacheUnbindActions(Type type)
         {
-            var ourType = GetType();
+            // Generate delegates to unbind fields
+            var actions = type.GetFields(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.DeclaredOnly)
+                              .Where(f => typeof(IUnbindable).IsAssignableFrom(f.FieldType))
+                              .Select(f => new Action<object>(target => ((IUnbindable)f.GetValue(target))?.UnbindAll()));
 
-            if (unbind_action_cache.TryGetValue(ourType, out _))
-                return;
-
-            // List containing all the delegates to perform the unbinds
-            var actions = new List<Action<object>>();
-
-            foreach (var type in ourType.EnumerateBaseTypes())
-            {
-                if (unbind_action_cache.TryGetValue(type, out var action))
-                {
-                    actions.Add(action);
-                    continue;
-                }
+            // Add parent type's action.
+            var baseType = type.BaseType;
+            if (baseType != null)
+                actions = actions.Append(unbind_action_cache.TryGetValue(baseType, out var action) ? action : cacheUnbindActions(baseType));
 
-                // Generate delegates to unbind fields
-                actions.AddRange(type.GetFields(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.DeclaredOnly)
-                                     .Where(f => typeof(IUnbindable).IsAssignableFrom(f.FieldType))
-                                     .Select(f => new Action<object>(target => ((IUnbindable)f.GetValue(target))?.UnbindAll())));
+            var actionList = actions.ToList();
 
-                // Delegates to unbind properties are intentionally not generated.
-                // Properties with backing fields (including automatic properties) will be picked up by the field unbind delegate generation,
-                // while ones without backing fields (like get-only properties that delegate to another drawable's bindable) should not be unbound here.
+            // Delegates to unbind properties are intentionally not generated.
+            // Properties with backing fields (including automatic properties) will be picked up by the field unbind delegate generation,
+            // while ones without backing fields (like get-only properties that delegate to another drawable's bindable) should not be unbound here.
 
-                unbind_action_cache[type] = target =>
+            Action<object> finalAction = target =>
+            {
+                foreach (var a in actionList)
                 {
-                    foreach (var a in actions)
+                    try
                     {
-                        try
-                        {
-                            a(target);
-                        }
-                        catch (Exception e)
-                        {
-                            Logger.Error(e, $"Failed to unbind a local bindable in {type.ReadableName()}");
-                        }
+                        a(target);
                     }
-                };
-            }
+                    catch (Exception e)
+                    {
+                        Logger.Error(e, $"Failed to unbind a local bindable in {type.ReadableName()}");
+                    }
+                }
+            };
+
+            unbind_action_cache[type] = finalAction;
+            return finalAction;
         }
 
         private bool unbindComplete;
@@ -285,7 +278,9 @@ private void load(IFrameBasedClock clock, IReadOnlyDependencyContainer dependenc
 
             InjectDependencies(dependencies);
 
-            cacheUnbindActions();
+            var type = GetType();
+            if (!unbind_action_cache.ContainsKey(type))
+                cacheUnbindActions(type);
 
             LoadAsyncComplete();
 
```